options - fix indentation and improve logic

+++ docs/developer/scripts/k3d-dev.sh
@@ -27,51 +27,55 @@ while [ -n "$1" ]; do # while loop starts
 	case "$1" in
 
 	-b) echo "-b option passed for big k3d cluster using M5 instance" 
-        BIG_INSTANCE=true
-        ;;
+      BIG_INSTANCE=true
+    ;;
 
 	-p) echo "-p option passed to cretate k3d cluster with private ip" 
-        PRIVATE_IP=true
-        ;;
+      PRIVATE_IP=true
+    ;;
 
 	-m) echo "-m option passed to install MetalLB" 
-        METAL_LB=true
-        ;; 
+      METAL_LB=true
+    ;; 
 
 	-d) echo "-d option passed to destroy the AWS resources"
-		AWSINSTANCEIDs=$(aws ec2 describe-instances \
-    		--output text \
-    		--query "Reservations[].Instances[].InstanceId" \
-    		--filters "Name=tag:Name,Values=${AWSUSERNAME}-dev" "Name=instance-state-name,Values=running" )
+	  AWSINSTANCEIDs=$( aws ec2 describe-instances \
+	    --output text \
+	    --query "Reservations[].Instances[].InstanceId" \
+	    --filters "Name=tag:Name,Values=${AWSUSERNAME}-dev" "Name=instance-state-name,Values=running" )
+      # If instance exists then terminate it 
+      if [[ ! -z "${AWSINSTANCEIDs}" ]]; then 
 		echo "aws instances being terminated: ${AWSINSTANCEIDs}"
+		# TODO: should we add a user confirmation prompt here for safety?
 		aws ec2 terminate-instances --instance-ids ${AWSINSTANCEIDs}
 		echo -n "waiting for instance termination..."
 		aws ec2 wait instance-terminated --instance-ids ${AWSINSTANCEIDs}
 		echo "done"
-		echo "SecurityGroup name to be deleted: ${SGname}"
-		aws ec2 delete-security-group --group-name=${SGname}
-		echo "KeyPair to be deleted: ${KeyName}"
-		aws ec2 delete-key-pair --key-name ${KeyName}
-		exit 0 
-        ;;
-    -h)
-        echo "Usage:"
-        echo "k3d-dev.sh -b -p -m -d -h"
-        echo ""
-        echo " -b   use BIG M5 instance. Default is t3.2xlarge"
-        echo " -p   use private IP for security group and k3d cluster"
-        echo " -m   create k3d cluster with metalLB"
-		echo " -d   destroy related AWS resources"
-        echo " -h   output help"
-        exit 0
-        ;;
+	  else
+	    echo "You had no running instances."
+	  fi
+	  echo "SecurityGroup name to be deleted: ${SGname}"
+	  aws ec2 delete-security-group --group-name=${SGname}
+	  echo "KeyPair to be deleted: ${KeyName}"
+	  aws ec2 delete-key-pair --key-name ${KeyName}
+	  exit 0 
+	;;
+
+    -h) echo "Usage:"
+      echo "k3d-dev.sh -b -p -m -d -h"
+      echo ""
+      echo " -b   use BIG M5 instance. Default is t3.2xlarge"
+      echo " -p   use private IP for security group and k3d cluster"
+      echo " -m   create k3d cluster with metalLB"
+	  echo " -d   destroy related AWS resources"
+      echo " -h   output help"
+      exit 0
+    ;;
 
 	*) echo "Option $1 not recognized" ;; # In case you typed a different option other than a,b,c
 
 	esac
-
 	shift
-
 done
